# src/main/java/org/quwerty/notepadserver/dto/UserAccessDTO.java
package org.quwerty.notepadserver.dto;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Builder
@Getter
@Setter(AccessLevel.PACKAGE)
@AllArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE)
public class UserAccessDTO {
    int userId;
    String accessType;
}